+++ R/cross_categorical.R
@@ -1,7 +1,7 @@
 
 #' @importFrom dplyr everything mutate select
 #' @importFrom forcats fct_drop
-#' @importFrom purrr map_df
+#' @importFrom tibble as_tibble
 #' @keywords internal
 #' @noRd
 cross_categorical=function(data_x, data_y, showNA, total, label, percent_digits, percent_pattern,
@@ -35,17 +35,16 @@ cross_categorical=function(data_x, data_y, showNA, total, label, percent_digits,
   }
 
   rtn = rtn %>%
-    mutate(.id=names(data_x), label=x_name) %>%
+    mutate(.id=names(data_x), label=unname(x_name)) %>%
     select(".id", "label", everything()) %>%
-    map_df(as.character)
+    as_tibble()
 
   rtn
 }
 
 
-#' @importFrom dplyr across bind_rows filter matches mutate select transmute
+#' @importFrom dplyr across bind_rows filter matches mutate select starts_with transmute
 #' @importFrom glue glue
-#' @importFrom purrr map_df
 #' @importFrom tibble tibble
 #' @importFrom tidyr replace_na
 #' @keywords internal
@@ -94,12 +93,12 @@ summarize_categorical_single = function(x, showNA, total, digits, percent_patter
     rtn = bind_rows(rtn, .total)
   }
 
-  rtn %>% map_df(as.character)
+  rtn
 }
 
 
 
-#' @importFrom dplyr across bind_rows everything filter left_join mutate select transmute
+#' @importFrom dplyr across bind_rows everything filter left_join mutate select starts_with transmute
 #' @importFrom forcats fct_na_value_to_level
 #' @importFrom glue glue
 #' @importFrom purrr reduce
@@ -306,7 +305,6 @@ get_percent_pattern = function(margin=c("row", "column", "cell", "none", "all"),
   rtn$body = glue("{{n}} ({x})")
 
   if(isTRUE(na)){
-    # browser()
     ppv = percent_pattern_variables()
     ppv = ppv$na %>% set_names(ppv$std)
     rtn = map(rtn, ~str_replace_all(.x, ppv))
@@ -336,7 +334,7 @@ getTable = function(x, by, type=c("n", "p_tot", "p_row", "p_col")){
 
 #' @keywords internal
 #' @noRd
-#' @importFrom dplyr across mutate starts_with
+#' @importFrom dplyr mutate
 getTableCI = function(x, digits, method="wilson"){
   x %>%
     mutate(
@@ -350,7 +348,7 @@ getTableCI = function(x, digits, method="wilson"){
 }
 
 #' @importFrom purrr map_lgl
-#' @importFrom stringr str_subset str_detect
+#' @importFrom stringr str_detect str_subset
 #' @keywords internal
 #' @noRd
 percent_pattern_contains = function(percent_pattern, needle){