+++ R/flextable.R
@@ -69,8 +69,13 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
                                                        effect="effect"),
                                    ...) {
   assert_class(x, "crosstable", .var.name=vname(x))
+  if(any(dim(x)==0)){
+    cli_abort("{.fun as_flextable} cannot apply to an empty crosstable.",
+              class="crosstable_af_empty_error")
+  }
 
   if(missing(keep_id)) keep_id = getOption("crosstable_keep_id", keep_id)
+  if(missing(by_header)) by_header = getOption("crosstable_by_header", by_header)
   if(missing(autofit)) autofit = getOption('crosstable_autofit', autofit)
   if(missing(compact)) compact = getOption('crosstable_compact', compact)
   if(missing(show_test_name)) show_test_name = getOption('crosstable_show_test_name', show_test_name)
@@ -108,8 +113,7 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
   has_by =  !is.null(by)
   if(has_by && is.null(by_label)) by_label=by
   if(isTRUE(by_header)) by_header=NULL
-  if(length(by_header)==1 && is.na(by_header)) by_header=FALSE
-  if(identical(by_header, "")) by_header=FALSE
+  if(!is.null(by_header) && all(is.na(by_header) | by_header=="", na.rm=TRUE)) by_header=FALSE
 
   inner_labels = attr(x, "inner_labels")
   if(!is.null(inner_labels)){
@@ -213,7 +217,7 @@ as_flextable.crosstable = function(x, keep_id=FALSE, by_header=NULL,
   } else if(n_levels>1) {
     if(!is.null(by_header)){
       cli_warn("by_header is ignored if the crosstable has several `by` stratum.",
-               class = "crosstable_asflex_byheader_multi")
+               class="crosstable_af_byheader_multi")
     }
     header_mapping = tibble(col_keys = names(x)) %>%
       separate(col_keys, into=paste0(".col_", seq(n_levels)),